requirements.txt
streamlit
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
whisper==1.1.10
requests==2.31.0
transformers==4.34.0
torch==2.0.1
torchaudio
sentence-transformers==2.2.2
pyaudio==0.2.11
sounddevice==0.4.6
soundfile==0.12.1
playsound==1.3.0
striprtf==0.0.25
beautifulsoup4==4.12.2
odfpy==1.4.1
pygame==2.5.0
psycopg2-binary
sqlalchemy
pydub
SpeechRecognition
